refactor: 💡 change directory making dummy contents methods
